# Optimize date function performance by optimizing locks in DateLUT