@liuneng1994 2024-04-10
In centos7.9, the lock here will cause serious concurrency problems, making the performance close to the performance of single-threaded serial. very strange.