@alexey-milovidov 2024-04-10
It will be much better not to call `instance` in a loop.