@liuneng1994 2024-04-11
![image](https://github.com/ClickHouse/ClickHouse/assets/16730247/8ee85f13-598a-4f73-8293-1491108c55d7)
I tried other solutions, but they didn't work. This shouldn't be the bottleneck. But I discovered the real reason for the poor concurrency performance of the date parsing function. In the figure, try catch should not be used within the loop for error handling. This will cause serious lock waiting. It seems that the implementation of catch itself has an internal synchronization mechanism. Currently I'm trying to solve this problem.